Move code to `inc/namespace.php`
